+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## [Unreleased]
 - change the position of the jEnv status bar, before lineSeparator and after position 
-- remove duplicate jEnv jdk. if jEnv jdk canonical path exists, delete all. if jEnv jdk home path has one more, delete them until one exists.
+- remove duplicate jEnv jdk. Canonical path exists, delete all. Multiple same home path, delete them until only one exists.
+- add `Invalid Jenv` separator in the pop menu of jEnv status bar, click item to remove this JDK.
 
 ### Fixes
 - jEnv status bar icon no change when the project not set up JDK